feat: Add vulnerability and license policy processing with strategy pattern
- Introduced new handlers for building comments related to vulnerability and license policy violations.
- Created a strategy pattern for processing vulnerability and license policies, allowing for cleaner and more maintainable code.
- Implemented `PolicyProcessor` interface and concrete implementations for vulnerability and license policies.
- Refactored existing logic to utilize the new strategy pattern for evaluating policies.
- Added unit tests for the new processing logic and ensured existing tests are updated accordingly.
- Removed redundant code and improved overall structure of the handlers.
